# Speedup 10:  25 % quicker _get_array_dicts and 10 % quicker _prep_data_for_correlation