--- a/CHANGES.md
+++ b/CHANGES.md
@@ -12,6 +12,8 @@
   - `_prep_data_for_correlation`: 3x speedup for filling NaN-traces in templates
   - New function ``quick_trace_select` for a very efficient selection of trace
     by seed ID without wildcards (4x speedup).
+* utils.correlate
+  - 25 % speedup for `_get_array_dicts` with quicker access to properties.
 * utils.catalog_to_dd._prepare_stream
   - Now more consistently slices templates to length = extract_len * samp_rate
     so that user receives less warnings about insufficient data.