--- a/eqcorrscan/utils/pre_processing.py
+++ b/eqcorrscan/utils/pre_processing.py
@@ -999,18 +999,20 @@ def _prep_data_for_correlation(stream, templates, template_names=None,
                 template.traces[idx] for idx in stream_trace_id_dict[seed_id]])
             if len(template_channel) <= channel_index:
                 # out_template[channel_number].data = nan_channel  # quicker:
-                out_template[channel_number].__dict__['data'] = copy.deepcopy(
-                    nan_channel)
-                out_template[channel_number].stats.__dict__['npts'] = \
-                    template_length
-                out_template[channel_number].stats.__dict__['starttime'] = \
-                    template_starttime
-                out_template[channel_number].stats.__dict__['endtime'] = \
-                    UTCDateTime(ns=int(
+                out_template.traces[channel_number].__dict__[
+                    'data'] = copy.deepcopy(nan_channel)
+                out_template.traces[channel_number].stats.__dict__[
+                    'npts'] = template_length
+                out_template.traces[channel_number].stats.__dict__[
+                    'starttime'] = template_starttime
+                out_template.traces[channel_number].stats.__dict__[
+                    'endtime'] = UTCDateTime(ns=int(
                         round(template_starttime.ns
                               + (template_length * samp_rate) * 1e9)))
             else:
-                out_template[channel_number] = template_channel[channel_index]
+                out_template.traces[channel_number] = template_channel.traces[
+                    channel_index]
+
         # If a template-trace matches a NaN-trace in the stream , then set
         # template-trace to NaN so that this trace does not appear in channel-
         # list of detections.