### What does this PR do?
Implements:
- 25 % speedup for `utils.correlate._get_array_dicts` (from 200 s to 154 seconds total time spent in my test)
- 10 % speedup for `utils.pre_processing._prep_data_for_correlation` (from 702 s to 640 s total time spent)
- (for comparison to the numbers above: 494 s were spent for correlating data in parallel on the GPU)

Main speedups with:
- direct operations on `UTCDateTime.__dict__['_UTCDateTime__ns']` (5x quicker for one +/- operation)
- index-based access to trace in stream with `stream.traces[j]` instead of `stream[j]` (4.5x quicker)

### Why was it initiated?  Any relevant Issues?
For big datasets run on powerful hardware, the cumulative time spent in `utils.correlate._get_array_dicts` and `utils.pre_processing._prep_data_for_correlation` makes up a considerable part (>50 %) of a detection run. This PR tries to optimize the functions with some simple changes; I think further speed ups would need changes that are not so trivial...

This PR is part of the list of speedups in https://github.com/eqcorrscan/EQcorrscan/issues/522

### PR Checklist
- [X] `develop` base branch selected?
- [X] This PR is not directly related to an existing issue (which has no PR yet).
- [] All tests still pass.
~- [ ] Any new features or fixed regressions are be covered via new tests.~
~- [ ] Any new or changed features have are fully documented.~
- [X] Significant changes have been added to `CHANGES.md`.
~- [ ] First time contributors have added your name to `CONTRIBUTORS.md`.~