@flixha 2023-01-10
Sorry for the mess with 2 commits with the same message; I thought I had reset / removed the first one locally before pushing. As it's so small I think it's easier to leave it than trying to completely remove that commit from remote; anyways it's fixed by the last commit.